requirements.txt
mysql-connector
Flask
requests
